20180412答辩版本，答辩时的bug原因：下载的个人页面模板所使用的框架需要联网。。

+++ src/main/webapp/WEB-INF/back/activity/list.jsp
@@ -101,7 +101,7 @@
                 <td>${Activity.overtime}</td>
                 <td>
 
-                    <a href="${pageContext.request.contextPath }/back/activityDelete.do?activityId=${Activity.activityId}" class="tablelink" >删除</a>
+                    <a href="${pageContext.request.contextPath }/back/activityDelete.do?activityId=${Activity.activityId}" class="tablelink" onClick="return confirm('确定删除?');">删除</a>
                 </td>
             </tr>
         </c:forEach>
